refactor(client): improves architecture by adding service and repository in each page directory
