+++ client/src/context/profile/auth.context.tsx
@@ -1,8 +1,7 @@
 import { createContext, ReactNode, useEffect, useState } from 'react';
-
 import { ProfileQuery } from 'graphql/__generated__/graphql';
 import { ApolloQueryResult, OperationVariables } from '@apollo/client';
-import { useProfile } from 'api/profile/use-profile';
+import { useProfile } from 'context/profile/profile.service';
 export interface AuthContextProps {
 	profile: ProfileQuery | null;
 	isLoading: boolean;
