--- client/src/graphql/profile/delete-account.mutation.ts
@@ -1,11 +1,0 @@
-import { gql } from '@apollo/client';
-
-export const DELETE_ACCOUNT_MUTATION = gql`
-	mutation DeleteAccount($password: String!) {
-		deleteMemberAccount(password: $password) {
-			id
-			email
-			username
-		}
-	}
-`;
