--- client/src/graphql/profile/update-email.mutation.ts
@@ -1,10 +1,0 @@
-import { gql } from '@apollo/client';
-
-export const UPDATE_EMAIL_MUTATION = gql`
-	mutation UpdateEmail($email: String!) {
-		updateMemberEmail(email: $email) {
-			email
-			id
-		}
-	}
-`;
