--- client/src/graphql/profile/update-password.mutation.ts
@@ -1,18 +1,0 @@
-import { gql } from '@apollo/client';
-
-export const UPDATE_PASSWORD_MUTATION = gql`
-	mutation UpdatePassword(
-		$newPassword: String!
-		$confirmedNewPassword: String!
-		$oldPassword: String!
-	) {
-		updateMemberPassword(
-			newPassword: $newPassword
-			confirmedNewPassword: $confirmedNewPassword
-			oldPassword: $oldPassword
-		) {
-			id
-			email
-		}
-	}
-`;
