--- client/src/graphql/profile/update-username.mutation.ts
@@ -1,10 +1,0 @@
-import { gql } from '@apollo/client';
-
-export const UPDATE_USERNAME_MUTATION = gql`
-	mutation UpdateUsername($username: String!) {
-		updateMemberUsername(username: $username) {
-			username
-			id
-		}
-	}
-`;
