+++ client/src/index.tsx
@@ -5,8 +5,8 @@ import 'styles/tailwind.css';
 import App from 'app';
 import reportWebVitals from 'reportWebVitals';
 import { ApolloClient, InMemoryCache, ApolloProvider } from '@apollo/client';
-import { AuthProvider } from 'context/auth.context';
-import { ProjectProvider } from 'context/project.context';
+import { AuthProvider } from 'context/profile/auth.context';
+import { ProjectProvider } from 'context/project/project.context';
 
 const client = new ApolloClient({
 	uri: '/',
