+++ client/src/pages/code/_components/file-explorer-panel.tsx
