+++ client/src/pages/code/code.container.tsx
@@ -1,10 +1,10 @@
 import Loader from 'components/svgs/loader';
-import ProjectContext from 'context/project.context';
 import { useAutoSaveProject, useGetProjectService } from 'pages/code/code.service';
 import { useContext } from 'react';
 import { useParams } from 'react-router-dom';
 import { ProjectContextData } from 'types/project';
 import { CodePage } from './code.page';
+import ProjectContext from 'context/project/project.context';
 
 export type Dependencies = Record<string, string>;
 export interface ProjectState {
