+++ client/src/pages/code/code.repository.ts
@@ -30,3 +30,28 @@ export const UPDATE_PROJECT_MUTATION = gql`
 		}
 	}
 `;
+
+export const GET_PROJECT_BY_ID_QUERY = gql`
+	query GetProjectById($projectId: String!) {
+		getProjectById(projectId: $projectId) {
+			owner {
+				id
+				sessions {
+					member {
+						sessions {
+							member {
+								id
+							}
+						}
+					}
+				}
+			}
+			name
+			isTemplate
+			id
+			isPublic
+			files
+			sandpackTemplate
+		}
+	}
+`;
