--- client/src/pages/create-project/helpers/getSandpackTemplateFromParam.ts
@@ -1,3 +1,0 @@
-// const getSandPackTemplateKey = (templateParam) => {
-
-// }
