+++ client/src/pages/home/_components/sign-out-button.tsx
