+++ client/src/pages/home/home.service.tsx
@@ -1,6 +1,6 @@
 import { useMutation } from '@apollo/client';
 import { SignOutMutation, SignOutMutationVariables } from 'graphql/__generated__/graphql';
-import { SIGN_OUT_MUTATION } from 'graphql/sign-out/sign-out.mutation';
+import { SIGN_OUT_MUTATION } from 'pages/home/home.repository';
 
 export const useSignOut = (onSignOutSuccess: () => Promise<void>) => {
 	const [signOut, { loading }] = useMutation<SignOutMutation, SignOutMutationVariables>(
