+++ client/src/pages/sign-up/sign-up.service.tsx
@@ -1,6 +1,6 @@
 import { useMutation } from '@apollo/client';
 import { SignUpMutation, SignUpMutationVariables } from 'graphql/__generated__/graphql';
-import { SIGN_UP_MUTATION } from 'graphql/signUp/sign-up.mutation';
+import { SIGN_UP_MUTATION } from 'pages/sign-up/sign-up.repository';
 
 export const useSignUp = () => {
 	const [signUp, { loading }] = useMutation<SignUpMutation, SignUpMutationVariables>(
